# openapi.json
{
  "components": {
    "schemas": {
...
      "EndpointResponse": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/WebhookState"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "event_types",
          "state",
          "created_at"
        ],
        "title": "EndpointResponse",
        "type": "object"
      },
      "EventTypesResponse": {
        "properties": {
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          }
        },
        "required": [
          "event_types"
        ],
        "title": "EventTypesResponse",
        "type": "object"
      },
      "FieldProblem": {
        "description": "One field-level validation message, as the 422 handler emits them.",
        "properties": {
          "field": {
            "description": "Dotted path including the source, e.g. body.title",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "FieldProblem",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 ``application/problem+json`` — **every** error response.\n\nDeclared as a model, and attached to the routers, for a reason that is easy to\nmiss: the exception handlers are invisible to FastAPI's schema generation, so\nwithout this the OpenAPI document would describe only the happy paths. An\nintegrator reading ``/docs`` would then have to *discover* the error shape by\ncausing errors, which is precisely the \"first thing an integrator hits\" §8.6\nis about. It also means the shape lands in API-5's snapshot, so changing it\nbecomes a visible contract change.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present on a 422; one entry per rejected field.",
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable, and names the next step.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the error class. **Match on this**, never on ``title`` — the title is human-readable prose and may be reworded.",
            "examples": [
              "https://relay.internal/problems/not_found"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "RegisterPayload": {
        "properties": {
          "event_types": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Event Types",
            "type": "array"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "event_types"
        ],
        "title": "RegisterPayload",
        "type": "object"
      },
      "RegisteredResponse": {
        "properties": {
          "endpoint": {
            "$ref": "#/components/schemas/EndpointResponse"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "signature_header": {
            "default": "X-Relay-Signature: sha256=HMAC(secret, timestamp + '.' + body)",
            "title": "Signature Header",
            "type": "string"
          }
        },
        "required": [
          "endpoint",
          "secret"
        ],
        "title": "RegisteredResponse",
        "type": "object"
      },
      "StatePayload": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/WebhookState"
          }
        },
        "required": [
          "state"
        ],
        "title": "StatePayload",
        "type": "object"
      },
...
      "WebhookState": {
        "enum": [
          "active",
          "paused",
          "disabled"
        ],
        "title": "WebhookState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Relay",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "operationId": "list_endpoints_api_v1_webhooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EndpointResponse"
                  },
                  "title": "Response List Endpoints Api V1 Webhooks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "List Endpoints",
        "tags": [
          "webhooks (v1)"
        ]
      },
      "post": {
        "description": "Register a destination. Refuses private, loopback and metadata targets (S-13).\n\nThe refusal checks the **resolved** addresses as well as the literal host, so\na name that resolves into private space is refused here rather than at the\nfirst delivery.",
        "operationId": "register_api_v1_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Register",
        "tags": [
          "webhooks (v1)"
        ]
      }
    },
...
    "/api/v1/webhooks/event-types": {
      "get": {
        "description": "The four events, so a subscriber does not have to hard-code strings.",
        "operationId": "event_types_api_v1_webhooks_event_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTypesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Event Types",
        "tags": [
          "webhooks (v1)"
        ]
      }
    },
    "/api/v1/webhooks/{endpoint_id}": {
      "delete": {
        "operationId": "delete_api_v1_webhooks__endpoint_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Delete",
        "tags": [
          "webhooks (v1)"
        ]
      }
    },
    "/api/v1/webhooks/{endpoint_id}/secret": {
      "post": {
        "description": "Mint a new secret. **The old one stops working immediately** — there is no\noverlap window, because we can only send one signature per delivery. Rotate,\nthen update the consumer.",
        "operationId": "rotate_api_v1_webhooks__endpoint_id__secret_post",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Rotate",
        "tags": [
          "webhooks (v1)"
        ]
      }
    },
    "/api/v1/webhooks/{endpoint_id}/state": {
      "put": {
        "description": "Pause or resume. Pausing **keeps** queued events, which is the whole\ndifference from deleting: a consumer being redeployed should not lose the\nevents that arrive during the window.",
        "operationId": "set_state_api_v1_webhooks__endpoint_id__state_put",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Endpoint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Refused — e.g. a tenant_id in the request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Out of scope for this token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource — or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "rev mismatch; the body carries the current rev"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed; see errors[]"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          }
        },
        "summary": "Set State",
        "tags": [
          "webhooks (v1)"
        ]
      }
    }
  }
}